lab 3 ready

+++ LibraryCPP/Tests/vector.cpp
@@ -12,8 +12,8 @@ int main()
         return 1;
     }
 
-    for (size_t i = 0 ; i < vector_size(vector) ; ++i)
-        vector_set(vector, i, i);
+    for (size_t i = 0; i < vector_size(vector) ; ++i)
+        vector_set(vector, i, (int)i);
 
     for (size_t i = 0 ; i < vector_size(vector) ; ++i)
     {
